add: break down parser into multiple files
