+++ codebank/headerbank/mpir_parser/mpir_parse.h
@@ -1,0 +1,14 @@
+/*
+ * GPL v3 - MPIR Compiler :: Copyright (C) 2023 Toby Benjamin Clark. All hail Gallaxhar!
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, of version 3 or later - See LICENSE for full terms of use.
+ */
+
+#ifndef MPIR_COMPILER_MPIR_PARSE_H
+#define MPIR_COMPILER_MPIR_PARSE_H
+
+#include "../../headerbank/mpir_parser/mpir_parser.h"
+#include "../../headerbank/mpir_parser/mpir_parser_utilities.h"
+#include "../../headerbank/mpir_parser/mpir_parse_funcdef.h"
+
+#endif //MPIR_COMPILER_MPIR_PARSE_H
