+++ codebank/sourcebank/mpir_parser/mpir_parser.c
@@ -109,156 +109,7 @@ void mpir_parser_free(mpir_parser* parser)
     free(parser);
 }
 
-struct mpir_identifier* parse_identifier(mpir_parser* psr)
-{
-    struct mpir_identifier node;
-    if((psr->peek(psr))->type != IDENTIFIER)
-    {
-        mpir_error("parse_function_declaration: expected function identifier got other.");
-        return NULL;
-    }
-    else wcscpy(node.data, (psr->get(psr))->lexeme);
-    return &node;
-}
-
-struct mpir_identifier* parse_returntype(mpir_parser* psr)
-{
-    struct mpir_type node;
-    if((psr->peek(psr))->type != IDENTIFIER)
-    {
-        mpir_error("parse_function_declaration: expected function returntype got other.");
-        return NULL;
-    }
-    else wcscpy(node.data, (psr->get(psr))->lexeme);
-    wprintf(L"Parsed Return Type of '%ls' \n", node.data);
-    return &node;
-}
-
-struct mpir_identifier* parse_type(mpir_parser* psr)
-{
-    struct mpir_type node;
-    if((psr->peek(psr))->type != IDENTIFIER)
-    {
-        mpir_error("parse_function_declaration: expected function type got other.");
-        return NULL;
-    }
-
-    else if((psr->peek(psr))->type == IDENTIFIER)
-    {
-        wcscpy(node.data, (psr->get(psr))->lexeme);
-        wprintf(L"Type Identifier '%ls' \n", node.data);
-        return &node;
-    }
-    else return NULL;
-}
-
-struct mpir_type** parse_inputs_internal(mpir_parser* psr, struct mpir_type** nodes, int node_index)
-{
-    /*
-     * Needs to dynamically read & process types i.e. Int, Int or Int, Float,
-     */
-
-    if((psr->peek(psr))->type != IDENTIFIER)
-    {
-        mpir_error("parse_function_declaration: expected function identifier got other.");
-        return NULL;
-    }
-
-    /* Allocate Memory for list of types */
-    nodes = realloc(nodes, sizeof(struct mpir_type*) * (node_index + 1));
-    nodes[0] = parse_type(psr);
-
-    if((psr->peek(psr))->type == keyword_comma)
-    {
-        /* The ',' token is voided, as it has no semantic integrity. */
-        (void)psr->get(psr);
-        node_index++;
-        parse_inputs_internal(psr, nodes, node_index);
-    }
-    else if(psr->peek(psr)->type == operator_arrow)
-    {
-        /* No more types */
-        return nodes;
-    }
-    else
-    {
-        mpir_error("Parser expected ', Type' or ->");
-    }
-}
-
-struct mpir_type** parse_inputs(mpir_parser* psr)
-{
-    struct mpir_type** nodes;
-    nodes = malloc(sizeof(struct mpir_type*));
-    return parse_inputs_internal(psr, nodes, 0);
-}
-
-struct mpir_function_declaration* parse_function_declaration(mpir_parser* psr)
-{
-    /*
-     * 'funcdef' identifier '::' function_io '\n'
-     */
-
-    /* Create Funcdef AST node & Consume 'fundef' */
-    struct mpir_function_declaration node;
-    /* Parsing */
-
-    /* Parse `funcdef */
-    if(psr->peek(psr)->type != keyword_funcdef) return NULL;
-    else if(psr->peek(psr)->type == keyword_funcdef) (void)psr->get(psr);
-    else return NULL;
-
-    /* Parse function identifier */
-    if(psr->peek(psr)->type == IDENTIFIER) node.identifier = psr->get(psr);
-    else return NULL;
-    if(node.identifier == NULL) return NULL;
-    wprintf(L"Function Identifier: '%ls'\n", node.identifier);
-
-    /* Parse I/O shield operator `::` */
-    if(psr->peek(psr)->type != double_colon) return NULL;
-    else if(psr->peek(psr)->type == double_colon) (void)psr->get(psr);
-    else return NULL;
-    printf("Parsed :: \n");
-
-    /* Parse return type */
-    if((node.inputs = parse_inputs(psr)) == NULL) return NULL;
-    if(!(psr->tryget(psr, operator_arrow))) return NULL;
-    if((node.return_type = parse_returntype(psr)) == NULL) return NULL;
-
-    /* Parse Newline */
-    if(psr->peek(psr)->type == NEWLINE) (void)psr->get(psr);
-    else return NULL;
-
-    /* Add Declaration Header to Program */
-    insert_at_end(psr->program, (union command_data){.function_declaration = &node});
-
-    return psr;
-}
-
-void mpir_parse(mpir_parser* parser)
-{
-    mpir_token_type next_type;
-    mpir_token* current_token = parser->get(parser);
-    while(current_token != NULL)
-    {
-        if(parser->peek(parser) != NULL)
-        {
-            next_type = (parser->peek(parser))->type;
-            if(next_type == keyword_funcdef)
-            {
-                printf("branching to parse_function_declaration!\n");
-                parse_function_declaration(parser);
-            }
-        }
-        else next_type = (mpir_token_type) NULL;
-
-        current_token = parser->get(parser);
-    }
-    return;
-}
-
-/* Implement all Parsers to produce AST */
-/*
+/* Implement all Parsers to produce AST
  * |-- Function Definition: password_check
 |   |
 |   |-- Identifier: password_check
